Update main_win on window changes to avoid pointing to closed windows.
Fixes #14.

For now, we add a buffer-local autocommand that simply updates the state
variable each time we the buffer enters new window.  This fixes the
problem of re-entering into a hidden buffer (e.g. when you close the
tab and then open it again).

Currently the event is not triggerred by `:split`, therefore if one
opens two windows for the same buffer, the messages will be displayed
relatively to the original window.