Don't use utf-8 filter on partial results.
Sometimes, if we get partial message, it might be cut in the middle
of the utf8 symbol, in which case the utf8 library complains that
the input is not well-formed.  However, as we are only want to remove
prefixes from the string, we do not really care about that.

Therefore use regex functions of string.