Use reified names when showing context.
Sometimes when the names of hidden variables coincide, Agda
disambiguates them, yet it sends both versions (originalName and
reifiedName) in the response.  We used original names, which
created ambiguous contexts such as:

  n : Nat
  n : Nat

but in the goals and subexpressions it always uses reified versions.

Now we show correct name.  Here is a small example:

  data Nat : Set where
    zero : Nat
    suc : Nat → Nat

  data X : Nat → Set where
    triv : ∀ {n} → X (suc n)

  foo : ∀ {n} → X n → X (suc n)
  foo {n} triv = ?

and \, shows:

  ────────────────────────────
  Goal Type : X (suc (suc n₁))
  ────────────────────────────
  n₁ : Nat
  n  : Nat